fix the walking animation
the WASD button triggers the walking animation, but when change the direction from left to right, it should stay at Idle state first. But the error made the animation played even if it is just a change in character direction. the fix help the character stay in idle state when the button been press 1 time and triggered the walking animation when long press the button.